\name{extract_coefficient_path}
\alias{extract_coefficient_path}
\title{Extract Coefficient Path}
\usage{
  extract_coefficient_path(cvg)
}
\arguments{
  \item{cvg}{cv.glmnet objects}
}
\value{
  long data.frame of coeffients, penalties, and deviance
  ratios
}
\description{
  Extract coefficients along the regularization path
}
\author{
  Erik Shilts
}

