\name{plot_cv_path}
\alias{plot_cv_path}
\title{Plot Cross Validation Path}
\usage{
  plot_cv_path(cvg, ...)
}
\arguments{
  \item{cvg}{cv.glmnet object}

  \item{\dots}{additional plotting arguments}
}
\value{
  ggplot2 grid object
}
\description{
  Plots the cross validation path as seen in
  plot.cv.glmnet.
}
\author{
  Erik Shilts
}

