\name{plot_logistic_response}
\alias{plot_logistic_response}
\title{Plot Logistic Response}
\usage{
  plot_logistic_response(x, y, cvg, ...)
}
\arguments{
  \item{x}{input training data}

  \item{y}{dependent response variable}

  \item{cvg}{cv.glmnet object}

  \item{\dots}{other graphical parameters}
}
\value{
  ggplot2 created grid object
}
\description{
  Plots the logistic reponse
}
\author{
  Erik Shilts
}

